Convert a list to AWS CLI args

dots2args(...)

list2args(l)

Arguments

l
A list of arguments to be mapped to AWS CLI args. #' @details Unnamed list entries are inserted into the system call as (positional) "<val>" arguments, i.e. with no argument-specifying hyphens. Named list entries are inserted into the system call with the "--<name> <val>" pattern. Any named list entries with a NULL value are inserted into the system call as "--<name>" ut with no additional value, i.e. those arguments typically specify named flags. The AWS CLI call's arguments are built in the order in which they appear in the list, so if an argument appears more than once in the list, there'll be multiple such args in the system call, too (in the same order in which they appear in the list).

Examples

list2args(list("unnamed", "named" = "value", "flag" = NULL))

#> [1] "unnamed" "--named" "value"   "--flag"